requirements.txt
uvicorn[standard]==0.18.3
gunicorn==20.1.0
Django==4.0
python-dotenv==0.20.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.26.5
django-filter==22.1
Pillow==9.2.0
redis==4.3.4
channels-redis==3.4.0
channels==3.0.4
aioredis==1.3.1
whitenoise==6.2.0